Verplaats computed_fields van BuikInWandLLM naar BuikInWand
