+++ src/tekstherkenning_ark/llm/gebrek_classificatie.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 
 from typing import ClassVar, Literal
 
-from pydantic import BaseModel, Field, computed_field
+from pydantic import BaseModel, Field
 
 from tekstherkenning_ark.enums import NietBeschikbaar
 from tekstherkenning_ark.llm.llmclassifier import LLMClassifier
@@ -303,48 +303,6 @@ of NietBeschikbaar.NIET_MEETBAAR als de waarde niet meetbaar is."""
         default=NietBeschikbaar.LEEG,
         description="Mate van uitbuiging in centimeters. Zoek naar 'uitbuiging' of 'buik'.",
     )
-
-    @computed_field
-    @property
-    def start_buik_van_startrak_m(self) -> float | NietBeschikbaar:
-        """Afstand van het startrak in meters. Wordt afgeleid van de start van de buik."""
-        if self.start_buik_van_startrak_m_ingevuld is not NietBeschikbaar.LEEG:
-            return self.start_buik_van_startrak_m_ingevuld
-        elif (
-            self.eind_buik_van_startrak_m_ingevuld is not NietBeschikbaar.LEEG
-            and self.lengte_buik_m_ingevuld is not NietBeschikbaar.LEEG
-        ):
-            return self.eind_buik_van_startrak_m_ingevuld - self.lengte_buik_m_ingevuld
-        else:
-            return NietBeschikbaar.LEEG
-
-    @computed_field
-    @property
-    def eind_buik_van_startrak_m(self) -> float | NietBeschikbaar:
-        """Afstand van het startrak in meters. Wordt afgeleid van het eind van de buik."""
-        if self.eind_buik_van_startrak_m_ingevuld is not NietBeschikbaar.LEEG:
-            return self.eind_buik_van_startrak_m_ingevuld
-        elif (
-            self.start_buik_van_startrak_m_ingevuld is not NietBeschikbaar.LEEG
-            and self.lengte_buik_m_ingevuld is not NietBeschikbaar.LEEG
-        ):
-            return self.start_buik_van_startrak_m_ingevuld + self.lengte_buik_m_ingevuld
-        else:
-            return NietBeschikbaar.LEEG
-
-    @computed_field
-    @property
-    def lengte_buik_m(self) -> float | NietBeschikbaar:
-        """Lengte van de buik in meters. Wordt afgeleid van de start en eind van de buik."""
-        if self.lengte_buik_m_ingevuld is not NietBeschikbaar.LEEG:
-            return self.lengte_buik_m_ingevuld
-        elif (
-            self.start_buik_van_startrak_m_ingevuld is not NietBeschikbaar.LEEG
-            and self.eind_buik_van_startrak_m_ingevuld is not NietBeschikbaar.LEEG
-        ):
-            return self.eind_buik_van_startrak_m_ingevuld - self.start_buik_van_startrak_m_ingevuld
-        else:
-            return NietBeschikbaar.LEEG
 
 
 class ScheefstandLLM(LLMClassifier):
